IISC/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
pydantic
jose
deepl
email-validator 
pymongo[srv]
redis
PyJWT
routes
tokenizers